Basic matlab wrapper for analysis working.
* check matlab/test_oas_setup.m for example on how to specify design variables

+++ matlab/mat2np.m
@@ -1,7 +1,7 @@
 function npary = mat2np(mat)
 
 % convert matlab matrix to python (Numpy) ndarray 
-sh = fliplr(size(mat));
+sh = int32(fliplr(size(mat)));
 mat2 = reshape(mat,1,numel(mat));  % [1, n] vector
 npary = py.numpy.array(mat2);
 npary = npary.reshape(sh).transpose();  % python ndarray
